zendesk/etl/requirements.txt
zenpy
google-cloud-storage
cryptography
google-cloud-bigquery
pandas
pandas-gbq

